pyproject.toml
[project]
name = "pnpq"
dynamic = ["version"]
description = "PnPQ is a python library package for controlling optical devices in quantum networks"
authors = [
    { name = "Amin Taherkhani", email = "amin@sfc.wide.ad.jp" }
]
dependencies = [
    "pyserial>=3.5",
    "types-pyserial>=3.5.0.20240826",
    "pint>=0.24.3",
    "structlog>=24.4.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "pnpq:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.8.0",
    "isort>=5.13.2",
    "mypy>=1.11.2",
    "pytest>=8.3.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pnpq"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "src/pnpq/__init__.py"

[tool.isort]
profile = "black"

[tool.mypy]
files = ["."]
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
